+++ src/core/destroy.js
