+++ src/utils/slugify.js
@@ -1,0 +1,3 @@
+export default function slugify(...strings) {
+  return strings.join(' ').replace(/[^A-Za-z0-9]+/g, '-');
+}
