+++ tests/helpers.js
@@ -1,0 +1,3 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
